Fix BlendShape deformation broken by incorrect index buffer writing
When the componentType is UnsignedByte, the data written to byteArray must be 1 byte, but BitConverter.GetBytes((byte)index) returns a 2-byte byte array, so the vertex index of the BlendShape was not read correctly and the deformation by the BlendShape was broken.

The componentType was determined using builder.BaseAccessorCount, but the indices, which are the Values of the internal Dictionary of builder, are not necessarily continuous values, and the indices with a vertex displacement of 0 by the BlendShape are excluded, so changed to get the maximum value of indices.