[exporter] disable emission when an emission mask is present in lilToon
When lilToon's emission is enabled and an emission mask is present, the emission should be restricted according to the mask. However, since MToon does not support this functionality, the material emits light regardless of the emission mask, causing an issue.

Additionally, since lilToon does not have an implementation for baking emission separately, emission will be uniformly disabled when an emission mask is present.