[exporter] fix issue where multiple PB colliders were not considered
Previously, cases where a single game object contained multiple PB colliders were not accounted for. This is necessary, for example, when combining multiple sphere colliders to approximate a capsule collider. This update adds support for such cases.

While, in principle, multiple PBs and constraints could also be supported, this commit does not address that.