[exporter] Improves matcap processing (#65)
When rim light is disabled and matcap is enabled, the rim light settings are overridden from MToon's implementation where matcap depends on rim light. Also, it was discovered that when the matcap color is pure white, the implementation was returning null, which differs from lilToon's original implementation, so this has been fixed to return the matcap texture.