[exporter] Fixed a bug where hidden bones are referenced
Likely due to the effects of GH-66, found a bug where bones from hidden SMRs are referenced when they should not be, causing `KeyNotFoundException` in the following location. This has been fixed to not add them when they are hidden.

```csharp
var joints = resolver.UniqueTransforms.Select(bone => _transformNodeIDs[bone]).ToList();
```