[exporter] Fixed a bug where NaN is output
There was a bug where `depthRatio` would receive NaN due to division by zero, and when serialized as JSON, NaN values are treated as strings, making the file unreadable as VRM (since the bug is limited to VRM spring bones, there's no issue when treated as glTF). This has been fixed to check if the value is zero and return 0 if it is zero.