[exporter] Fixed a bug with unintended dependency on VRChat SDK extension method (#99)
It was discovered that `UnityEngine.UIElements.VisualElement.SetVisible` is implemented as an extension method in the VRChat SDK, and the method is not implemented in the minimum required VRChat SDK version, causing a compilation failure. Since there should be no dependency on the VRChat SDK, this has been fixed to directly implement the original implementation.

This fixes the issue in GH-98.